requirements.txt
fastapi
uvicorn
python-dotenv
httpx
openai>=1.0.0
google-generativeai
supabase
pydantic
absl-py>=2.3.0
aiohappyeyeballs>=2.6.1
aiohttp>=3.12.13
aiosignal>=1.3.2
annotated-types>=0.7.0
anyascii>=0.3.2
async-timeout>=5.0.1
attrs>=25.3.0
audioread>=3.0.1
babel>=2.17.0
bangla>=0.0.5
blinker>=1.9.0
blis>=1.2.1
bnnumerizer>=0.0.2
bnunicodenormalizer>=0.1.7
catalogue>=2.0.10
certifi>=2025.6.15
cffi>=1.17.1
charset-normalizer>=3.4.2
click>=8.2.1
cloudpathlib>=0.21.1
confection>=0.1.5
contourpy>=1.2.1
coqpit>=0.0.17
cycler>=0.12.1
cymem>=2.0.11
Cython>=3.1.2
dateparser>=1.1.8
decorator>=5.2.1
docopt>=0.6.2
einops>=0.8.1
encodec>=0.1.1
filelock>=3.18.0
Flask>=3.1.1
fonttools>=4.58.4
frozenlist>=1.7.0
fsspec>=2025.5.1
g2pkk>=0.1.2
grpcio>=1.73.1
gruut>=2.2.3
gruut-ipa>=0.13.0
gruut_lang_de>=2.0.1
gruut_lang_en>=2.0.1
gruut_lang_es>=2.0.1
gruut_lang_fr>=2.0.2
hangul-romanize>=0.1.0
hf-xet>=1.1.5
huggingface-hub>=0.33.1
idna>=3.10
inflect>=7.5.0
itsdangerous>=2.2.0
jamo>=0.4.1
jieba>=0.42.1
Jinja2>=3.1.6
joblib>=1.5.1
jsonlines>=1.2.0
kiwisolver>=1.4.8
langcodes>=3.5.0
language_data>=1.3.0
lazy_loader>=0.4
librosa>=0.10.0
llvmlite>=0.43.0
marisa-trie>=1.2.1
Markdown>=3.8.2
markdown-it-py>=3.0.0
MarkupSafe>=3.0.2
matplotlib>=3.8.4
mdurl>=0.1.2
more-itertools>=10.7.0
mpmath>=1.3.0
msgpack>=1.1.1
multidict>=6.5.1
murmurhash>=1.0.13
networkx>=2.8.8
nltk>=3.9.1
num2words>=0.5.14
numba>=0.60.0
numpy>=1.22.0
nvidia-cublas-cu12>=12.6.4.1
nvidia-cuda-cupti-cu12>=12.6.80
nvidia-cuda-nvrtc-cu12>=12.6.77
nvidia-cuda-runtime-cu12>=12.6.77
nvidia-cudnn-cu12>=9.5.1.17
nvidia-cufft-cu12>=11.3.0.4
nvidia-cufile-cu12>=1.11.1.6
nvidia-curand-cu12>=10.3.7.77
nvidia-cusolver-cu12>=11.7.1.2
nvidia-cusparse-cu12>=12.5.4.2
nvidia-cusparselt-cu12>=0.6.3
nvidia-nccl-cu12>=2.26.2
nvidia-nvjitlink-cu12>=12.6.85
nvidia-nvtx-cu12>=12.6.77
packaging>=25.0
pandas>=1.5.3
pillow>=11.2.1
platformdirs>=4.3.8
pooch>=1.8.2
preshed>=3.0.10
propcache>=0.3.2
protobuf>=6.31.1
psutil>=7.0.0
pycparser>=2.22
Pygments>=2.19.2
pynndescent>=0.5.13
pyparsing>=3.2.3
pypinyin>=0.54.0
pysbd>=0.3.4
python-crfsuite>=0.9.11
python-dateutil>=2.9.0.post0
pytz>=2025.2
PyYAML>=6.0.2
regex>=2024.11.6
requests>=2.32.4
rich>=14.0.0
safetensors>=0.5.3
scikit-learn>=1.7.0
scipy>=1.11.4
shellingham>=1.5.4
six>=1.17.0
smart-open>=7.1.0
sounddevice>=0.5.2
soundfile>=0.13.1
soxr>=0.5.0.post1
spacy>=3.8.7
spacy-legacy>=3.0.12
spacy-loggers>=1.0.5
srsly>=2.5.1
SudachiDict-core>=20250515
SudachiPy>=0.6.10
sympy>=1.14.0
tensorboard>=2.19.0
tensorboard-data-server>=0.7.2
thinc>=8.3.4
threadpoolctl>=3.6.0
tiktoken>=0.9.0
tokenizers>=0.21.2
torch>=2.7.1
torchaudio>=2.7.1
tqdm>=4.67.1
trainer>=0.0.36
transformers>=4.53.0
triton>=3.3.1
TTS>=0.22.0
typeguard>=4.4.4
typer>=0.16.0
typing-inspection>=0.4.1
typing_extensions>=4.14.0
tzlocal>=5.3.1
umap-learn>=0.5.7
Unidecode>=1.4.0
urllib3>=2.5.0
wasabi>=1.1.3
weasel>=0.4.1
Werkzeug>=3.1.3
wrapt>=1.17.2
yarl>=1.20.1
